requirements.txt
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.3.0
scipy>=1.6.0

